feat: :sparkles: adds methods to create/edit sections and departments fga-eps-mds/2021.1-Oraculo#146
